Add files via upload
updated the css file for new pages such as map and log in 
created the map.html
updated the previous login.html
created a map.js file to help make the map pretty 
used an API from leafletjs.com
IMPORTANT!
i have edited only the login.html file  for the log in function removed the original nav bar and added my own nav bar. removed a line that was <script src = (cant remember what it was but i think it had something to with a style sheet and shouldnt affect code)></script>
connected a script.js file to the login.html. removed all the previous css and created new css design to match the rest of the code.
END OF IMPORTANT!
Bishr can you test the login functionality with this new style and let me know if it still works or if not what errors occur